fix dtype

+++ tests_tf/test_attacks.py
@@ -911,8 +911,8 @@ class TestLogitSpaceProjectedGradientAscent(CleverHansTest):
         levels = 10
         x = np.random.rand(10, 32, 32, 3)
         y = np.ones(10)
-        x_t = tf.constant(x, tf.int32)
-        y_t = tf.one_hot(tf.constant(y), depth=10)
+        x_t = tf.constant(x, tf.float32)
+        y_t = tf.one_hot(tf.constant(y, dtype=tf.int32), depth=10)
         x_thermometer = discretize_uniform(
             x_t, levels, thermometer=True)
         x_adv = self.attack.generate(x, eps=eps, levels=levels, steps=steps,
